requirements.txt
fastapi
uvicorn[standard]
requests
python-dotenv
pydantic
pytest
scikit-learn 
numpy        